fix(ct): match dir in --dir option against bare tests in apps first
Prior to this change:
In umbrella projects when --dir apps/subapp is used, the dir is
treated as an extra source dir for root app (but not subapp) when
root app is ordered before other sub-apps in the apps list.

After this change:
It first tires to meatch against all apps' bare tests.
only after failed to match against all apps, it tries to match
it as root apps's extra.

This should avoid copying the whole apps dir into
_build/test/lib/rootapp/