fix: partial revert 80671df57a06915a23493b9fc721d1ae8f493964 for elixir
In the reverted commit, the app vsn setting after compile was
removed because it was believed that app vsn is set during app
discovery stage and hence no need to set again after compile

Elixir apps do not have a .app.src file.
This leads to an issue for elixir dependencies (by rebar_mix plugin):
the app is not set even if the .app file is generated after compile.