fix(git): Unshallow repo before --reference clone
Prior to this change, the reference repo is unshallowed after
cache auto fill.
In case the unshallow command is aborted, it will fail the
reference clones.
This commit makes rebar3 try to unshallow the repo before
performing --reference clone.

Also, if reference clone failed, it fallbacks to the regular clone.